requirements.txt
python==3.8.8
pytorch==1.8.1
pandas==1.2.4
scikit-learn==0.23.2
numpy==1.20.1
matplotlib==3.3.4
scipy==1.6.2